WebGateway/wg-svc/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
DateTime==4.9
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-decouple==3.6
pytz==2022.7
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zope.interface==5.5.2
gunicorn==20.1.0
Flask-Cors==3.0.10
pymongo==4.0.1
Flask-PyMongo==2.3.0
